% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEP functions.R
\name{ASE_detection}
\alias{ASE_detection}
\title{Perform gene-level ASE detection across genes}
\usage{
ASE_detection(dat_all, phased = FALSE, adaptive = TRUE,
  n_resample = 10^6, parallel = FALSE, n_core = 1,
  save_out = FALSE, name_out = NULL)
}
\arguments{
\item{dat_all:}{bulk RNA-seq data. Must contain variables with names: "gene", "id", "ref", "total" \cr
* "gene: character, gene name;
* "id": character, individual identifier;
* "ref": numeric, the snp-level read counts for the reference allele if the haplotype phase of the data is unknown, and the snp-level read counts for allele aligned on paternal/maternal haplotype if haplotype phase is known;
* "total": numeric, snp-level total read counts for both alleles; \cr}

\item{phased:}{a logical value indicates whether the RNA-seq data is properly phased with respect to the haplotype or not. Default is FALSE}

\item{n_resample:}{a numeric value indicates the maximum number of resamplings performed to obtain estimated p-value. Default is 10^6}

\item{adaptive:}{a logical value indicates whether the resampling is done through an adaptive procedure or not. Default is TRUE \cr
By adaptive, it means first do 1000 resamplings, if the estimated p-value < 0.1, increase the number of resampling, by a factor of 10, to 10^4.
if then the estimated p-value < 0.01, increase the number of resampling again, by a factor of 10, to 10^5.
The procedure continuous until reaches the maximum number of resampling.}

\item{parallel:}{a logical value indicates whether do parallel computing for the resampling precedure or not. Default is FALSE}

\item{n_core:}{a numeric value indicates number of clusters used for the parallel computing. Used only when parallel is set to TRUE. Default is 1}

\item{save_out:}{a logical value indicates whether to write out the result for each gene stepwisely to a txt file. Default is FALSE}

\item{name_out:}{a character string indicates the output file name when save_out is set to TRUE, with the format of "XXX.txt"}
}
\value{
A matrix with three columns:
      * "gene": character, gene name;
      * "test statistic": numeric, the test statistics of ASE effect;
      * "p-value": the estimated p-value of the test statistic;
}
\description{
This function is used to detect ASE genes given bulk RNA-seq data that may contain multiple genes.
}
