% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEP functions.R
\name{differential_ASE_detection}
\alias{differential_ASE_detection}
\title{Perform gene-level differential ASE analysis in the population across genes}
\usage{
differential_ASE_detection(
  dat_all,
  phased = FALSE,
  varList = NULL,
  adaptive = TRUE,
  n_resample = 10^6,
  parallel = FALSE,
  n_core = 1,
  save_out = FALSE,
  name_out = NULL
)
}
\arguments{
\item{dat_all:}{bulk RNA-seq data. Must contain variables: \cr
\itemize{
   \item `gene`: character, gene name;
   \item `id`: character, individual identifier;
   \item `snp`: character, the name/chromosome location of the heterzygous genetic variants;
   \item `ref`: numeric, the snp-level read counts for the reference allele if the haplotype phase of the data is unknown, and the snp-level read counts for allele aligned on the same paternal/maternal haplotype for both conditions if haplotype phase is known;
   \item `total`: numeric, snp-level total read counts for both alleles;
   \item `group`: character, the condition each RNA-seq sample is obtained from (i.e., pre- vs post-treatment);
   \item `ref_condition`: character, the condition used as the reference for pseudo haplotype phasing;
}}

\item{phased:}{a logical value indicates whether the haplotype phase of the data is known or not. Default is FALSE}

\item{varList:}{a character string specifies fomula of covariates that users want to adjusted in the model. An example could be "`var1`+`var2`". Default is NULL}

\item{n_resample:}{a numeric value indicates the maximum number of resamplings performed to obtain estimated p-value. Default is 10^6}

\item{adaptive:}{a logical value indicates whether the resampling is done through an adaptive procedure or not. Default is TRUE \cr
By adaptive, it means first do 1000 resamplings, if the estimated p-value < 0.1, increase the number of resampling, by a factor of 10, to 10^4.
if then the estimated p-value < 0.01, increase the number of resampling again, by a factor of 10, to 10^5.
The procedure continuous until reaches the maximum number of resampling.}

\item{parallel:}{a logical value indicates whether do parallel computing for the resampling precedure or not. Default is FALSE}

\item{n_core:}{a numeric value indicates number of clusters used for parallel computing when parameter "parallel" is set to TRUE. Default is 1}

\item{save_out:}{a logical value indicates whether to write out the result for each gene stepwisely to a txt file. Default is FALSE}

\item{name_out:}{a character string indicates the output file name when save_out is set to TRUE, with the format of "XXX.txt"}
}
\value{
A matrix with three columns:
      \itemize{
         \item `gene`: character, gene name;
         \item `p-value`: the estimated p-value of the likelihood ratio test (LRT) statistic;
      }
}
\description{
This function is used to detect differential ASE genes between two conditions given bulk RNA-seq data that may contain multiple genes.
}
