% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEP functions.R
\name{modelFit}
\alias{modelFit}
\title{Model Fitting}
\usage{
modelFit(dat_phase, n_condition = "one", resampled_data = FALSE)
}
\arguments{
\item{dat_phase:}{Psudo-phased bulk RNA-seq data of a given gene}

\item{n_condition:}{a character string indicates whether to perform one condition analysis or two conditions analysis. Possible values are "one" or "two". Default is "one"}

\item{resampled_data:}{a logical value indicates whether the data analyzed is obtained from resampling or not. Used only for two conditions analysis. Default is FALSE}
}
\value{
\itemize{
               \item One condition analysis: likelihood ratio test (LRT) statistic;\cr
               \item Two conditions analysis: \cr
                 - original data: estimated ASE effect in the pooled sample & likelihood ratio test (LRT) statistic; \cr
                 - resampled data: likelihood ratio test (LRT) statistic;\cr
   }
}
\description{
This function is used to fit the generalized linear mixed-effect model through nonparametric maximum likelihood estimation for a given gene.
}
