% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEP functions.R
\name{phasing}
\alias{phasing}
\title{Haplotype Pseudo Alignment}
\usage{
phasing(dat, phased = FALSE, n_condition = "one")
}
\arguments{
\item{dat:}{bulk RNA-seq dataset of a given gene. Must contain variables with names: \cr
* One condition analysis: "id", "ref", "total". \cr
  - "id": character, individual identifier;
  - "ref": numeric, the snp-level read counts for the reference allele if the haplotype phase of the data is unknown, and the snp-level read counts for allele aligned on paternal/maternal haplotype if haplotype phase is known;
  - "total": numeric, snp-level total read counts for both alleles;
* Two condition analysis: "id", "snp", "ref", "total", "group", "ref_condition". \cr
  - "snp": character, the name/chromosome location of the heterzygous genetic variants;
  - "group": character, the condition each RNA-seq sample is obtained from (i.e., pre- vs post-treatment);
  - "ref_condition": character, the condition used as the reference for pseudo haplotype phasing;}

\item{phased:}{a logical value indicates whether the RNA-seq data is properly phased with respect to the haplotype or not. Default is FALSE}

\item{n_condition:}{a character string indicates whether the RNA-seq data contains data from only one condition or two conditions (i.e. normal vs diseased). Possible values are "one" or "two". Default is "one"}
}
\value{
The psudo-phased data for the input RNA-seq data, where column "major" indicates the read counts for major alleles aligned across individuals
}
\description{
This function is used to obtain the pseudo haplotype phase of the RNA-seq data for a given gene, and align the major alleles across individuals.
}
