% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEP functions.R
\name{plot_ASE}
\alias{plot_ASE}
\title{Boxplot of the estimated SNP-level ASE}
\usage{
plot_ASE(dat, phased = FALSE)
}
\arguments{
\item{dat:}{bulk RNA-seq data of a given gene. Must contain variables: \cr
\itemize{
   \item `gene`: character, gene name;
   \item `id`: character, individual identifier;
   \item `snp`: character, the name/chromosome location of the heterzygous genetic variants;
   \item `ref`: numeric, the snp-level read counts for the reference allele if the haplotype phase of the data is unknown, and the snp-level read counts for allele aligned on the same paternal/maternal haplotype for both conditions if haplotype phase is known;
   \item `total`: numeric, snp-level total read counts for both alleles;
}}

\item{phased:}{a logical value indicates whether the haplotype phase of the data is known or not. Default is FALSE}
}
\value{
The boxplot (ggplot2::geom_boxplot) of the estimated SNP-level ASE across individuals
}
\description{
This function is used to showed the estimated SNP-level ASE, i.e. major allele proportion,
for one condition sample across all individuals and SNPs after haplotype phasing.
The individuals are aligned based on an increasing trend of their median ASE across all snps for a given gene.
}
