% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEP functions.R
\name{two_conditions_analysis_Gene}
\alias{two_conditions_analysis_Gene}
\title{Perform gene-level differential ASE analysis}
\usage{
two_conditions_analysis_Gene(dat, phased = FALSE, adaptive = TRUE,
  n_resample = 10^6, parallel = FALSE, n_core = 1)
}
\arguments{
\item{dat:}{bulk RNA-seq data of a given gene. Must contain variables with names: "id", "snp", "ref", "total", "group", "ref_condition" \cr
* "id": character, individual identifier;
* "snp": character, the name/chromosome location of the heterzygous genetic variants;
* "ref": numeric, the snp-level read counts for the reference allele if the haplotype phase of the data is unknown, and the snp-level read counts for allele aligned on paternal/maternal haplotype if haplotype phase is known;
* "total": numeric, snp-level total read counts for both alleles;
* "group": character, the condition each RNA-seq sample is obtained from (i.e., pre- vs post-treatment);
* "ref_condition": character, the condition used as the reference for pseudo haplotype phasing;}

\item{phased:}{a logical value indicates whether the RNA-seq data is properly phased with respect to the haplotype or not. Default is FALSE}

\item{n_resample:}{a numeric value indicates the maximum number of resamplings performed to obtain estimated p-value. Default is 10^6}

\item{adaptive:}{a logical value indicates whether the resampling is done through an adaptive procedure or not. Default is TRUE \cr
By adaptive, it means first do 1000 resamplings, if the estimated p-value < 0.1, increase the number of resampling, by a factor of 10, to 10^4.
if then the estimated p-value < 0.01, increase the number of resampling again, by a factor of 10, to 10^5.
The procedure continuous until reaches the maximum number of resampling. \cr}

\item{parallel:}{a logical value indicates whether do parallel computing for the resampling precedure or not. Default is FALSE}

\item{n_core:}{a numeric value indicates number of clusters used for the parallel computing. Used only when parallel is set to TRUE. Default is 1}
}
\value{
The test statistic and its estimated p-value from resampling
}
\description{
This function is used to perform differential ASE detection for two conditions analysis of a given gene.
}
