% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEP functions.R
\name{modelFit}
\alias{modelFit}
\title{Fit Generalized Linear Mixed-Effect Model}
\usage{
modelFit(dat_phase, n_condition = "one", resampled_data = FALSE)
}
\arguments{
\item{dat_phase:}{Psudo-phased bulk RNA-seq data of a given gene}

\item{n_condition:}{a character string indicates whether to perform one condition analysis or two conditions analysis. Possible values are "one" or "two". Default is "one"}

\item{resampled_data:}{a logical value indicates whether the data analyzed is obtained from resampling or not. Used for two conditions analysis only. Default is FALSE}
}
\value{
One of: \cr
   * the test statistic for allele specific expression (ASE) detection - when modeling the original/resampled data under one condition analysis \cr
   * the test statistic for differential ASE detection - when modeling the resampled data under two conditions analysis \cr
   * the test statistic for differential ASE detection & the estimated ASE effect in the pooled sample - when modeling the original data under two conditions analysis \cr
}
\description{
This function is used to fit the generalized linear mixed-effect model through nonparametric maximum likelihood estimation for a given gene.
}
