% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEP functions.R
\name{one_condition_analysis_Gene}
\alias{one_condition_analysis_Gene}
\title{Perform ASE analysis in the population for a given gene}
\usage{
one_condition_analysis_Gene(
  dat,
  phased = FALSE,
  n_resample = 10^6,
  adaptive = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{dat:}{bulk RNA-seq data of a given gene. Must contain variables:
\itemize{
     \item `id`: character, individual identifier;
     \item `ref`: numeric, the snp-level read counts for the reference allele if the haplotype phase of the data is unknown, and the snp-level read counts for allele aligned on paternal/maternal haplotype if haplotype phase is known;
     \item `total`: numeric, snp-level total read counts for both alleles;
}}

\item{phased:}{a logical value indicates whether the haplotype phase of the data is known or not. Default is FALSE}

\item{n_resample:}{a numeric value indicates the maximum number of resamplings performed to obtain estimated p-value. Default is 10^6}

\item{adaptive:}{a logical value indicates whether the resampling is done through an adaptive procedure or not. Default is TRUE \cr
By adaptive, it means first do 1000 resamplings, if the estimated p-value < 0.1, increase the number of resampling, by a factor of 10, to 10^4.
if then the estimated p-value < 0.01, increase the number of resampling again, by a factor of 10, to 10^5.
The procedure continuous until reaches the maximum number of resampling. \cr}

\item{parallel:}{a logical value indicates whether do parallel computing for the resampling precedure or not. Default is FALSE}
}
\value{
A vector with two elements:
      \itemize{
         \item `LRT statistic`: numeric, the likelihood ratio test (LRT) statistics of ASE effect;
         \item `p-value`: the estimated p-value of the LRT statistic;
      }
}
\description{
This function is used to perform ASE detection for one condition analysis of a given gene.
}
