% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASEP functions.R
\name{plot_ASE_diff}
\alias{plot_ASE_diff}
\title{Plot of the estimated SNP-level ASE difference}
\usage{
plot_ASE_diff(dat, phased = FALSE, minu_ref = TRUE)
}
\arguments{
\item{dat:}{bulk RNA-seq data of a given gene. Must contain variables with names: \cr
* "gene": character, gene name;
* "id": character, individual identifier;
* "snp": character, the name/chromosome location of the heterzygous genetic variants;
* "ref": numeric, the snp-level read counts for the reference allele if the haplotype phase of the data is unknown, and the snp-level read counts for allele aligned on paternal/maternal haplotype if haplotype phase is known;
* "total": numeric, snp-level total read counts for both alleles;
* "group": character, the condition each RNA-seq sample is obtained from (i.e., pre- vs post-treatment);
* "ref_condition": character, the condition used as the reference for pseudo haplotype phasing; \cr}

\item{phased:}{a logical value indicates whether the RNA-seq data is properly phased with respect to the haplotype or not. Default is FALSE}

\item{minu_ref:}{a logical value indicates when calculating the difference, whether the condition treated as the minuend, i.e. the difference is calculated as estimated ASE for minuend_condition minus the other condition, is the same as "ref_condition". Default is TRUE}
}
\value{
The boxplot showed the estimated SNP-level ASE difference across individuals
}
\description{
This function is used to showed the estimated SNP-level ASE difference, i.e. major allele proportion difference,
between two condition samples across all individuals and SNPs after haplotype phasing.
The individuals are aligned based on an increasing trend of their median ASE differences across all snps for a given gene.
}
